requirements.txt
streamlit
google-generativeai
reportlab
requests
folium
streamlit-folium
geopy
